+++ Catalog.Host/Controllers/ManufacturerController.cs
@@ -2,6 +2,7 @@
 using Catalog.Host.Models.Dtos;
 using Catalog.Host.Models.Requests;
 using Catalog.Host.Models.Responses;
+using Catalog.Host.Models.Validators;
 using Catalog.Host.Services.Interfaces;
 using Infrastructure;
 using Microsoft.AspNetCore.Mvc;
@@ -22,22 +23,39 @@ public class ManufacturerController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> Add(string brand, string country)
+    [ProducesResponseType(typeof(AddCarResponse<int?>), (int)HttpStatusCode.OK)]
+    public async Task<IActionResult> Add([FromBody] CreateManufacturerRequest request)
     {
-        var result = await _brandService.Add(brand, country);
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var result = await _brandService.Add(request.ManufacturerName, request.ManufacturerCountry);
         return Ok(new AddCarResponse<int?>() { Id = result });
     }
 
     [HttpDelete]
     public IActionResult Delete(int itemId)
     {
+        if (itemId < 0)
+        {
+            return BadRequest("Id cannot be <0");
+        }
+
         var result = _brandService.Delete(itemId);
         return Ok();
     }
 
     [HttpPut]
-    public async Task<IActionResult> Put(ManufacturerDto item, int itemToUpdate)
+    public async Task<IActionResult> Put([FromBody] ManufacturerDto item, int itemToUpdate)
     {
+        var validationResult = await new ManufacturerDtoValidator().ValidateAsync(item);
+        if (!validationResult.IsValid)
+        {
+            return BadRequest(validationResult.Errors);
+        }
+
         var result = await _brandService.Put(item, itemToUpdate);
         return Ok();
     }
